Refactor LLVM export into focused modules
